requirements.txt
openai 
python-pptx
streamlit
dotenv